+++ scone/cs/PieceWiseLinearFunction.cpp
@@ -13,7 +13,7 @@ namespace scone
 	PieceWiseLinearFunction::PieceWiseLinearFunction( const PropNode& props, opt::ParamSet& par )
 	{
 		size_t control_points;
-		INIT_PROPERTY( props, control_points, 0u );
+		INIT_PROPERTY( props, control_points, size_t( 0 ) );
 		INIT_PROPERTY( props, flat_extrapolation, false );
 
 		for ( size_t cpidx = 0; cpidx < control_points; ++cpidx )
